+++ libs/hooks/src/lib/queries/useAdminLiveShoppingList.tsx
@@ -1,23 +1,7 @@
-import { LiveShopping } from '@prisma/client';
+import { LiveShoppingWithGoods } from '@project-lc/shared-types';
 import { AxiosError } from 'axios';
 import { useQuery, UseQueryResult } from 'react-query';
-import { BroadcasterDTO } from '@project-lc/shared-types';
 import axios from '../../axios';
-
-export interface LiveShoppingWithGoods extends LiveShopping {
-  goods: {
-    goods_name: string;
-    summary: string;
-  };
-  seller: {
-    sellerShop: {
-      sellerEmail: string;
-      shopName: string;
-    };
-  };
-  broadcaster: BroadcasterDTO;
-  liveShoppingVideo: { youtubeUrl: string };
-}
 
 export const getAdminLiveShoppingList = async (
   liveShoppingId?: string | null,
