+++ libs/hooks/src/lib/queries/useBroadcasterLiveShoppingList.tsx
@@ -1,23 +1,8 @@
-import { LiveShopping } from '@prisma/client';
 import { AxiosError } from 'axios';
-import { useQuery, UseQueryResult, UseQueryOptions } from 'react-query';
-import { BroadcasterDTO } from '@project-lc/shared-types';
+import { useQuery, UseQueryResult } from 'react-query';
+import { LiveShoppingWithGoods } from '@project-lc/shared-types';
 import axios from '../../axios';
 
-interface LiveShoppingWithGoods extends LiveShopping {
-  goods: {
-    goods_name: string;
-    summary: string;
-  };
-  seller: {
-    sellerShop: {
-      sellerEmail: string;
-      shopName: string;
-    };
-  };
-  broadcaster: BroadcasterDTO;
-  liveShoppingVideo: { youtubeUrl: string };
-}
 // return type any로 안하면 리액트 테이블에서 컬럼 린트 에러 발생
 export const getBroadcasterLiveShoppingList = async (
   broadcasterId: number | undefined,
