+++ libs/shared-types/src/lib/res-types/liveShoppingWithGoods.res.ts
@@ -1,0 +1,17 @@
+import { LiveShopping } from '@prisma/client';
+import { BroadcasterDTO } from '../dto/broadcaster.dto';
+
+export interface LiveShoppingWithGoods extends LiveShopping {
+  goods: {
+    goods_name: string;
+    summary: string;
+  };
+  seller: {
+    sellerShop: {
+      sellerId: number;
+      shopName: string;
+    };
+  };
+  broadcaster: BroadcasterDTO;
+  liveShoppingVideo: { youtubeUrl: string };
+}
